pyproject.toml
[project]
name = "2024-umiusi-scripts"
version = "0.1.0"
description = "umiusi便利スクリプトたち"
authors = [
    { name = "H1rono", email = "hronok66@gmail.com" }
]
dependencies = [
    "pyserial-asyncio>=0.6",
    "PyQt6>=6.7.1",
    "matplotlib>=3.9.1",
    "numpy>=2.0.1",
    "aiochannel>=1.2.1",
    "typing-extensions>=4.12.2",
    "setuptools>=71.1.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[project.scripts]
nucleo-read-forever = "nucleo_communicate.scripts:run_read_forever"
nucleo-plot-flex1 = "nucleo_communicate.scripts:flex1_animation"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.rye]
managed = true
dev-dependencies = []
